Convert indexing to functional style / stream processing.
The method that takes a text, splits it into words, stems each
word, converts it to lower case, and then counts the words --
this method seemed to be a good candidate for stream processing.

It's very succinct, which is nice.  Could be a bear to debug, though.

Unfortunately, the fallout is that the word counts are now Long
instead of Integer.  That was the easiest to get out of the stream.
Maybe it's possible to get Integer counts out of the stream.
Anyway, right now, we just assume the words don't occur too often,
and just assume that we can downcast the term frequency to an integer.

+++ src/main/java/de/emptydomain/springwicket/DocumentsController.java
@@ -104,7 +104,7 @@ public class DocumentsController {
 	{
 		model.addAttribute("query", query);
 
-		Map<String, Integer> queryWords = TextHelper.indexText(query);
+		Map<String, Long> queryWords = TextHelper.indexText(query);
 
 		Map<String, Long> documentFrequency = new HashMap<>();
 		for (String w : queryWords.keySet()) {
